Refactor user interface components for improved consistency and readability
- Updated FieldMetricsWindow to streamline rendering of neural field metrics, consolidating card rendering logic and improving layout.
- Enhanced LogWindow and PlotControlWindow to adopt a consistent title bar style with icons.
- Modified NodeGraphWindow to replace the node icon for better representation.
- Adjusted SimulationWindow to integrate FieldMetricsWindow for monitoring neural fields, enhancing the overall user experience.
- Refined StaticLayoutWindow to utilize consistent background colors for child elements, improving visual coherence.
