+++ dynamic-neural-field-composer/include/user_interface/plot_control_window.h
@@ -16,6 +16,7 @@
 #endif
 
 extern ImFont* g_BlackLargeFont;
+extern ImFont* g_MediumIconsFont;
 
 namespace dnf_composer::user_interface
 {
